Remove xslt=false GET parameter
This didn't seem to work with the current approach of using XSLT for
parameters. This was also not documented. And it interferred with a
future plan to cache the results of XSLT transformations (as it would be
possible that some were executed with transform=true and others with
transform=false, and the caching would have to take account of that.)