11.0.0: Remove facility to configure request log expiry
As the Service Portal user cannot see any request log entries older than
the month they're in or 30 days back, it does not make sense to keep
request log entries older than that.

This was previously configurable, but it made no sense to use a longer
value (as the user could not see those not-expired request log entries
over 30 days), nor a shorter value (the UI would display date options
which contained no data). As there was only one value that made sense,
that is now hardcoded into the software and is one less thing for the
deployment engineer to consider.